+++ src/sdk/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Support for using PnP Core SDK with other cloud environments such as the US government clouds (GCC, GCC High and DoD), Germany and China clouds [jansenbe - Bert Jansen]
 - Added support to for using `new Guid()` in PnP Core SDK LINQ queries #633 [s-KaiNet - Sergei Sergeev]
 - Added support for list web hooks #630 [s-KaiNet - Sergei Sergeev]
+- Added `EnsureSiteAssetsLibrary` method on `IList` [jansenbe - Bert Jansen]
 
 ### Changed